AE-1780: Derived param/tests migration and clean-up
